pop up de confirmacion en eliminar rol

+++ frontend-application/src/app/adm-usuarios/roles/editar-rol/editar-rol.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { Rol } from '../../../model/rol.model';
 import { RolService } from './../../../services/rol.service';
+import { ConfirmacionPopupComponent } from '../../confirmacion-popup/confirmacion-popup.component';
+import { MatDialog } from '@angular/material';
 
 @Component({
   selector: 'app-editar-rol',
@@ -9,7 +11,8 @@ import { RolService } from './../../../services/rol.service';
 })
 export class EditarRolComponent implements OnInit {
 
-  constructor(private _rolService: RolService) { }
+  constructor(private _rolService: RolService,
+              private dialog: MatDialog) { }
 
   roles: Rol[];
 
@@ -28,6 +31,16 @@ export class EditarRolComponent implements OnInit {
   }
 
   deleteRol(rol: Rol){
-    this._rolService.deleteRol(rol).subscribe();
+    const dialogRef = this.dialog.open(ConfirmacionPopupComponent,{
+      width: '40%',
+      data: { mensaje: "Desea eliminar el rol?"}
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+
+        if (result && result == "true"){
+          this._rolService.deleteRol(rol).subscribe();
+        }
+    });
   }
 }
